Create README.md
A QR code (Quick Response code) is a type of matrix barcode that contains data readable by a QR scanner or smartphone. In this project, we will create a QR code generator using Python. The generator will take a string input (like a URL or text) and output a QR code image representing that input.